fix: storerate -> store request

+++ storerate/src/main/java/com/backend/storerate/dataobject/request/ChangeArticleGradeRequest.java
@@ -10,6 +10,6 @@ import javax.validation.constraints.NotNull;
 public class ChangeArticleGradeRequest {
     @NotNull
     private Long articleId;
-    private Long newGrade;
-    private Long customerNumber;
+    private Float newGrade;
+    private int customerNumber;
 }
